requirements.txt
streamlit
feedparser
pyyaml
requests
beautifulsoup4
python-dateutil